backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-CORS
PyJWT
python-dotenv
SQLAlchemy
werkzeug